// generated/src/app/api-models/stage-psclm-lens.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StagePsclmLens {
  batchId: string;
  transactionId: number;
  odSpherePm: number;
  odSphere: number;
  odSpherePb: number;
  odCylinderPm: number;
  odCylinder: number;
  odAxis: number;
  odPrism1Amt: number;
  odPrism1Dir: number;
  odPrism2Amt: number;
  odPrism2Dir: number;
  odAdd: number;
  osSpherePm: number;
  osSphere: number;
  osSpherePb: number;
  osCylinderPm: number;
  osCylinder: number;
  osAxis: number;
  osPrism1Amt: number;
  osPrism1Dir: number;
  osPrism2Amt: number;
  osPrism2Dir: number;
  osAdd: number;
  segType: string;
  insertDatetime: Date;
}